Struct aws_sdk_mediaconnect::client::fluent_builders::UpdateFlowOutput

pub struct UpdateFlowOutput<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateFlowOutput<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<UpdateFlowOutputOutput, SdkError<UpdateFlowOutputError>> where
R::Policy: SmithyRetryPolicy<UpdateFlowOutputInputOperationOutputAlias, UpdateFlowOutputOutput, UpdateFlowOutputError, UpdateFlowOutputInputOperationRetryAlias>,

pub fn cidr_allow_list(self, inp: impl Into<String>) -> Self

Appends an item to CidrAllowList.

To override the contents of this collection use set_cidr_allow_list. The range of IP addresses that should be allowed to initiate output requests to this flow. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

pub fn set_cidr_allow_list(self, input: Option<Vec<String>>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

A description of the output. This description appears only on the AWS Elemental MediaConnect console and will not be seen by the end user.

pub fn set_description(self, input: Option<String>) -> Self

pub fn destination(self, inp: impl Into<String>) -> Self

The IP address where you want to send the output.

pub fn set_destination(self, input: Option<String>) -> Self

pub fn encryption(self, inp: UpdateEncryption) -> Self

The type of key used for the encryption. If no keyType is provided, the service will use the default setting (static-key).

pub fn set_encryption(self, input: Option<UpdateEncryption>) -> Self

pub fn flow_arn(self, inp: impl Into<String>) -> Self

The flow that is associated with the output that you want to update.

pub fn set_flow_arn(self, input: Option<String>) -> Self

pub fn max_latency(self, inp: i32) -> Self

The maximum latency in milliseconds for Zixi-based streams.

pub fn set_max_latency(self, input: Option<i32>) -> Self

pub fn media_stream_output_configurations(
self,
inp: impl Into<MediaStreamOutputConfigurationRequest>
) -> Self

Appends an item to MediaStreamOutputConfigurations.

To override the contents of this collection use set_media_stream_output_configurations. The media streams that are associated with the output, and the parameters for those associations.

pub fn set_media_stream_output_configurations(
self,
input: Option<Vec<MediaStreamOutputConfigurationRequest>>
) -> Self

pub fn min_latency(self, inp: i32) -> Self

The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

pub fn set_min_latency(self, input: Option<i32>) -> Self

pub fn output_arn(self, inp: impl Into<String>) -> Self

The ARN of the output that you want to update.

pub fn set_output_arn(self, input: Option<String>) -> Self

pub fn port(self, inp: i32) -> Self

The port to use when content is distributed to this output.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn protocol(self, inp: Protocol) -> Self

The protocol to use for the output.

pub fn set_protocol(self, input: Option<Protocol>) -> Self

pub fn remote_id(self, inp: impl Into<String>) -> Self

The remote ID for the Zixi-pull stream.

pub fn set_remote_id(self, input: Option<String>) -> Self

pub fn smoothing_latency(self, inp: i32) -> Self

The smoothing latency in milliseconds for RIST, RTP, and RTP-FEC streams.

pub fn set_smoothing_latency(self, input: Option<i32>) -> Self

pub fn stream_id(self, inp: impl Into<String>) -> Self

The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

pub fn set_stream_id(self, input: Option<String>) -> Self

pub fn vpc_interface_attachment(self, inp: VpcInterfaceAttachment) -> Self

The name of the VPC interface attachment to use for this output.

pub fn set_vpc_interface_attachment(
self,
input: Option<VpcInterfaceAttachment>
) -> Self