Module aws_sdk_mediaconnect::error

Modules

add_flow_outputs420_exception

See AddFlowOutputs420Exception

bad_request_exception

See BadRequestException

create_flow420_exception

See CreateFlow420Exception

forbidden_exception

See ForbiddenException

grant_flow_entitlements420_exception

See GrantFlowEntitlements420Exception

internal_server_error_exception

See InternalServerErrorException

not_found_exception

See NotFoundException

service_unavailable_exception

See ServiceUnavailableException

too_many_requests_exception

See TooManyRequestsException

Structs

AddFlowMediaStreamsError

AddFlowOutputs420Exception

Exception raised by AWS Elemental MediaConnect. See the error message and documentation for the operation for more information on the cause of this exception.

AddFlowOutputsError

AddFlowSourcesError

AddFlowVpcInterfacesError

BadRequestException

Exception raised by AWS Elemental MediaConnect. See the error message and documentation for the operation for more information on the cause of this exception.

CreateFlow420Exception

Exception raised by AWS Elemental MediaConnect. See the error message and documentation for the operation for more information on the cause of this exception.

CreateFlowError

DeleteFlowError

DescribeFlowError

DescribeOfferingError

DescribeReservationError

ForbiddenException

Exception raised by AWS Elemental MediaConnect. See the error message and documentation for the operation for more information on the cause of this exception.

GrantFlowEntitlements420Exception

Exception raised by AWS Elemental MediaConnect. See the error message and documentation for the operation for more information on the cause of this exception.

GrantFlowEntitlementsError

InternalServerErrorException

Exception raised by AWS Elemental MediaConnect. See the error message and documentation for the operation for more information on the cause of this exception.

ListEntitlementsError

ListFlowsError

ListOfferingsError

ListReservationsError

ListTagsForResourceError

NotFoundException

Exception raised by AWS Elemental MediaConnect. See the error message and documentation for the operation for more information on the cause of this exception.

PurchaseOfferingError

RemoveFlowMediaStreamError

RemoveFlowOutputError

RemoveFlowSourceError

RemoveFlowVpcInterfaceError

RevokeFlowEntitlementError

ServiceUnavailableException

Exception raised by AWS Elemental MediaConnect. See the error message and documentation for the operation for more information on the cause of this exception.

StartFlowError

StopFlowError

TagResourceError

TooManyRequestsException

Exception raised by AWS Elemental MediaConnect. See the error message and documentation for the operation for more information on the cause of this exception.

UntagResourceError

UpdateFlowEntitlementError

UpdateFlowError

UpdateFlowMediaStreamError

UpdateFlowOutputError

UpdateFlowSourceError

Enums

AddFlowMediaStreamsErrorKind

AddFlowOutputsErrorKind

AddFlowSourcesErrorKind

AddFlowVpcInterfacesErrorKind

CreateFlowErrorKind

DeleteFlowErrorKind

DescribeFlowErrorKind

DescribeOfferingErrorKind

DescribeReservationErrorKind

GrantFlowEntitlementsErrorKind

ListEntitlementsErrorKind

ListFlowsErrorKind

ListOfferingsErrorKind

ListReservationsErrorKind

ListTagsForResourceErrorKind

PurchaseOfferingErrorKind

RemoveFlowMediaStreamErrorKind

RemoveFlowOutputErrorKind

RemoveFlowSourceErrorKind

RemoveFlowVpcInterfaceErrorKind

RevokeFlowEntitlementErrorKind

StartFlowErrorKind

StopFlowErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateFlowEntitlementErrorKind

UpdateFlowErrorKind

UpdateFlowMediaStreamErrorKind

UpdateFlowOutputErrorKind

UpdateFlowSourceErrorKind