Module aws_sdk_mediaconnect::input

Modules

add_flow_media_streams_input

See AddFlowMediaStreamsInput

add_flow_outputs_input

See AddFlowOutputsInput

add_flow_sources_input

See AddFlowSourcesInput

add_flow_vpc_interfaces_input

See AddFlowVpcInterfacesInput

create_flow_input

See CreateFlowInput

delete_flow_input

See DeleteFlowInput

describe_flow_input

See DescribeFlowInput

describe_offering_input

See DescribeOfferingInput

describe_reservation_input

See DescribeReservationInput

grant_flow_entitlements_input

See GrantFlowEntitlementsInput

list_entitlements_input

See ListEntitlementsInput

list_flows_input

See ListFlowsInput

list_offerings_input

See ListOfferingsInput

list_reservations_input

See ListReservationsInput

list_tags_for_resource_input

See ListTagsForResourceInput

purchase_offering_input

See PurchaseOfferingInput

remove_flow_media_stream_input

See RemoveFlowMediaStreamInput

remove_flow_output_input

See RemoveFlowOutputInput

remove_flow_source_input

See RemoveFlowSourceInput

remove_flow_vpc_interface_input

See RemoveFlowVpcInterfaceInput

revoke_flow_entitlement_input

See RevokeFlowEntitlementInput

start_flow_input

See StartFlowInput

stop_flow_input

See StopFlowInput

tag_resource_input

See TagResourceInput

untag_resource_input

See UntagResourceInput

update_flow_entitlement_input

See UpdateFlowEntitlementInput

update_flow_input

See UpdateFlowInput

update_flow_media_stream_input

See UpdateFlowMediaStreamInput

update_flow_output_input

See UpdateFlowOutputInput

update_flow_source_input

See UpdateFlowSourceInput

Structs

AddFlowMediaStreamsInput

A request to add media streams to the flow.

AddFlowOutputsInput

A request to add outputs to the specified flow.

AddFlowSourcesInput

A request to add sources to the flow.

AddFlowVpcInterfacesInput

A request to add VPC interfaces to the flow.

CreateFlowInput

Creates a new flow. The request must include one source. The request optionally can include outputs (up to 50) and entitlements (up to 50).

DeleteFlowInput

DescribeFlowInput

DescribeOfferingInput

DescribeReservationInput

GrantFlowEntitlementsInput

A request to grant entitlements on a flow.

ListEntitlementsInput

ListFlowsInput

ListOfferingsInput

ListReservationsInput

ListTagsForResourceInput

PurchaseOfferingInput

A request to purchase a offering.

RemoveFlowMediaStreamInput

RemoveFlowOutputInput

RemoveFlowSourceInput

RemoveFlowVpcInterfaceInput

RevokeFlowEntitlementInput

StartFlowInput

StopFlowInput

TagResourceInput

The tags to add to the resource. A tag is an array of key-value pairs. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

UntagResourceInput

UpdateFlowEntitlementInput

The entitlement fields that you want to update.

UpdateFlowInput

A request to update flow.

UpdateFlowMediaStreamInput

The fields that you want to update in the media stream.

UpdateFlowOutputInput

The fields that you want to update in the output.

UpdateFlowSourceInput

A request to update the source of a flow.