Module aws_sdk_mediaconnect::model

Modules

add_media_stream_request

See AddMediaStreamRequest

add_output_request

See AddOutputRequest

destination_configuration

See DestinationConfiguration

destination_configuration_request

See DestinationConfigurationRequest

encoding_parameters

See EncodingParameters

encoding_parameters_request

See EncodingParametersRequest

encryption

See Encryption

entitlement

See Entitlement

failover_config

See FailoverConfig

flow

See Flow

fmtp

See Fmtp

fmtp_request

See FmtpRequest

grant_entitlement_request

See GrantEntitlementRequest

input_configuration

See InputConfiguration

input_configuration_request

See InputConfigurationRequest

interface

See Interface

interface_request

See InterfaceRequest

listed_entitlement

See ListedEntitlement

listed_flow

See ListedFlow

media_stream

See MediaStream

media_stream_attributes

See MediaStreamAttributes

media_stream_attributes_request

See MediaStreamAttributesRequest

media_stream_output_configuration

See MediaStreamOutputConfiguration

media_stream_output_configuration_request

See MediaStreamOutputConfigurationRequest

media_stream_source_configuration

See MediaStreamSourceConfiguration

media_stream_source_configuration_request

See MediaStreamSourceConfigurationRequest

messages

See Messages

offering

See Offering

output

See Output

reservation

See Reservation

resource_specification

See ResourceSpecification

set_source_request

See SetSourceRequest

source

See Source

source_priority

See SourcePriority

transport

See Transport

update_encryption

See UpdateEncryption

update_failover_config

See UpdateFailoverConfig

vpc_interface

See VpcInterface

vpc_interface_attachment

See VpcInterfaceAttachment

vpc_interface_request

See VpcInterfaceRequest

Structs

AddMediaStreamRequest

The media stream that you want to add to the flow.

AddOutputRequest

The output that you want to add to this flow.

DestinationConfiguration

The transport parameters that are associated with an outbound media stream.

DestinationConfigurationRequest

The transport parameters that you want to associate with an outbound media stream.

EncodingParameters

A collection of parameters that determine how MediaConnect will convert the content. These fields only apply to outputs on flows that have a CDI source.

EncodingParametersRequest

A collection of parameters that determine how MediaConnect will convert the content. These fields only apply to outputs on flows that have a CDI source.

Encryption

Information about the encryption of the flow.

Entitlement

The settings for a flow entitlement.

FailoverConfig

The settings for source failover

Flow

The settings for a flow, including its source, outputs, and entitlements.

Fmtp

FMTP

FmtpRequest

The settings that you want to use to define the media stream.

GrantEntitlementRequest

The entitlements that you want to grant on a flow.

InputConfiguration

The transport parameters that are associated with an incoming media stream.

InputConfigurationRequest

The transport parameters that you want to associate with an incoming media stream.

Interface

The VPC interface that is used for the media stream associated with the source or output.

InterfaceRequest

The VPC interface that you want to designate where the media stream is coming from or going to.

ListedEntitlement

An entitlement that has been granted to you from other AWS accounts.

ListedFlow

Provides a summary of a flow, including its ARN, Availability Zone, and source type.

MediaStream

A single track or stream of media that contains video, audio, or ancillary data. After you add a media stream to a flow, you can associate it with sources and outputs on that flow, as long as they use the CDI protocol or the ST 2110 JPEG XS protocol. Each source or output can consist of one or many media streams.

MediaStreamAttributes

Attributes that are related to the media stream.

MediaStreamAttributesRequest

Attributes that are related to the media stream.

MediaStreamOutputConfiguration

The media stream that is associated with the output, and the parameters for that association.

MediaStreamOutputConfigurationRequest

The media stream that you want to associate with the output, and the parameters for that association.

MediaStreamSourceConfiguration

The media stream that is associated with the source, and the parameters for that association.

MediaStreamSourceConfigurationRequest

The definition of a media stream that you want to associate with the source.

Messages

Messages that provide the state of the flow.

Offering

A savings plan that reserves a certain amount of outbound bandwidth usage at a discounted rate each month over a period of time.

Output

The settings for an output.

Reservation

A pricing agreement for a discounted rate for a specific outbound bandwidth that your MediaConnect account will use each month over a specific time period. The discounted rate in the reservation applies to outbound bandwidth for all flows from your account until your account reaches the amount of bandwidth in your reservation. If you use more outbound bandwidth than the agreed upon amount in a single month, the overage is charged at the on-demand rate.

ResourceSpecification

A definition of what is being billed for, including the type and amount.

SetSourceRequest

The settings for the source of the flow.

Source

The settings for the source of the flow.

SourcePriority

The priority you want to assign to a source. You can have a primary stream and a backup stream or two equally prioritized streams.

Transport

Attributes related to the transport stream that are used in a source or output.

UpdateEncryption

Information about the encryption of the flow.

UpdateFailoverConfig

The settings for source failover

VpcInterface

The settings for a VPC Source.

VpcInterfaceAttachment

The settings for attaching a VPC interface to an output.

VpcInterfaceRequest

Desired VPC Interface for a Flow

Enums

Algorithm

Colorimetry

DurationUnits

EncoderProfile

EncodingName

EntitlementStatus

FailoverMode

KeyType

MediaStreamType

NetworkInterfaceType

PriceUnits

Protocol

Range

ReservationState

ResourceType

ScanMode

SourceType

State

Status

Tcs