Module aws_sdk_mediaconnect::operation

Structs

AddFlowMediaStreams

Adds media streams to an existing flow. After you add a media stream to a flow, you can associate it with a source and/or an output that uses the ST 2110 JPEG XS or CDI protocol.

AddFlowOutputs

Adds outputs to an existing flow. You can create up to 50 outputs per flow.

AddFlowSources

Adds Sources to flow

AddFlowVpcInterfaces

Adds VPC interfaces to flow

CreateFlow

Creates a new flow. The request must include one source. The request optionally can include outputs (up to 50) and entitlements (up to 50).

DeleteFlow

Deletes a flow. Before you can delete a flow, you must stop the flow.

DescribeFlow

Displays the details of a flow. The response includes the flow ARN, name, and Availability Zone, as well as details about the source, outputs, and entitlements.

DescribeOffering

Displays the details of an offering. The response includes the offering description, duration, outbound bandwidth, price, and Amazon Resource Name (ARN).

DescribeReservation

Displays the details of a reservation. The response includes the reservation name, state, start date and time, and the details of the offering that make up the rest of the reservation (such as price, duration, and outbound bandwidth).

GrantFlowEntitlements

Grants entitlements to an existing flow.

ListEntitlements

Displays a list of all entitlements that have been granted to this account. This request returns 20 results per page.

ListFlows

Displays a list of flows that are associated with this account. This request returns a paginated result.

ListOfferings

Displays a list of all offerings that are available to this account in the current AWS Region. If you have an active reservation (which means you’ve purchased an offering that has already started and hasn’t expired yet), your account isn’t eligible for other offerings.

ListReservations

Displays a list of all reservations that have been purchased by this account in the current AWS Region. This list includes all reservations in all states (such as active and expired).

ListTagsForResource

List all tags on an AWS Elemental MediaConnect resource

PurchaseOffering

Submits a request to purchase an offering. If you already have an active reservation, you can’t purchase another offering.

RemoveFlowMediaStream

Removes a media stream from a flow. This action is only available if the media stream is not associated with a source or output.

RemoveFlowOutput

Removes an output from an existing flow. This request can be made only on an output that does not have an entitlement associated with it. If the output has an entitlement, you must revoke the entitlement instead. When an entitlement is revoked from a flow, the service automatically removes the associated output.

RemoveFlowSource

Removes a source from an existing flow. This request can be made only if there is more than one source on the flow.

RemoveFlowVpcInterface

Removes a VPC Interface from an existing flow. This request can be made only on a VPC interface that does not have a Source or Output associated with it. If the VPC interface is referenced by a Source or Output, you must first delete or update the Source or Output to no longer reference the VPC interface.

RevokeFlowEntitlement

Revokes an entitlement from a flow. Once an entitlement is revoked, the content becomes unavailable to the subscriber and the associated output is removed.

StartFlow

Starts a flow.

StopFlow

Stops a flow.

TagResource

Associates the specified tags to a resource with the specified resourceArn. If existing tags on a resource are not specified in the request parameters, they are not changed. When a resource is deleted, the tags associated with that resource are deleted as well.

UntagResource

Deletes specified tags from a resource.

UpdateFlow

Updates flow

UpdateFlowEntitlement

You can change an entitlement’s description, subscribers, and encryption. If you change the subscribers, the service will remove the outputs that are are used by the subscribers that are removed.

UpdateFlowMediaStream

Updates an existing media stream.

UpdateFlowOutput

Updates an existing flow output.

UpdateFlowSource

Updates the source of a flow.